Brings in master's #11937 "Unify types with Geth", which flips the
foundational BlockHeader.Number and StateId.BlockNumber from long to
ulong (PreGenesis/Sync sentinels move from -1/long.MinValue to the top
of the ulong range, ulong.MaxValue/ulong.MaxValue-1).

Migrated the flat persisted-snapshot lineage to ulong block numbers to
match:

- Config (CompactSize/MaxReorgDepth/MinReorgDepth/LongFinalityMaxReorgDepth/
  PersistedSnapshotMaxCompactSize and the cache budgets) -> ulong.
- CompactionSchedule: two's-complement lowest-bit (~x+1), overflow-guarded
  NextFullCompactionAfter, PreGenesis treated as block 0 so the first
  finalized-persistence boundary still engages at genesis; offset stays
  long on disk (RLP back-compat).
- PersistenceManager: PreGenesis guards on the depth/seed/while comparisons
  (these compile but wrap silently without the guard); depth uses addition
  so no SaturatingSub needed.
- SnapshotRepository: a signed Height() helper for the backward-walk
  ordering comparisons — reinterpreting the block number as long maps the
  top-of-range sentinels (and an underflowing min bound) back to the
  pre-ulong signed order, so the walks behave exactly as before.
  Subtractions stay ulong (modular: x - PreGenesis == x + 1).
- Catalog on-disk key/value stays Int64; PreGenesis round-trips as -1.

Full solution builds clean; Nethermind.State.Flat.Test passes (777 passed,
4 pre-existing skips).

+++ src/Nethermind/Nethermind.Benchmark/State/WriteBatchBenchmark.cs
@@ -52,7 +52,7 @@ public class WriteBatchBenchmark
 
         int totalAccountCount = 0;
 
-        for (int block = 0; block < SnapshotCount; block++)
+        for (ulong block = 0; block < SnapshotCount; block++)
         {
             int accountCount = 500;
             int storageAccountCount = 10;
